various debugging

+++ crypto/aes.hh
@@ -4,6 +4,7 @@
 #include <assert.h>
 #include <openssl/aes.h>
 #include <vector>
+#include <cstring>
 
 class AES {
  public:
@@ -21,9 +22,10 @@ class AES {
         AES_decrypt(ctext, ptext, &dec);
     }
 
-    static const size_t blocksize = 16;
+    static const unsigned int blocksize = 16;
 
  private:
     AES_KEY enc;
     AES_KEY dec;
 };
+
